# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Api documentation Att Edwin Mogollon",
        "version": "1.0.0"
    },
    "paths": {
        "/timeStore": {
            "post": {
                "tags": [
                    "Respuestas"
                ],
                "summary": "Almacena el tiempo por pagina de un cliente",
                "description": "Almacena el tiempo por pagina de un cliente",
                "operationId": "4742bd9e5db2f9efa0e08bd00aa4366f",
                "parameters": [
                    {
                        "name": "Authorization_",
                        "in": "header",
                        "description": "Bearer token for authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Cliente login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "page": {
                                        "type": "string"
                                    },
                                    "survey": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "time": "234",
                                    "page": "1"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Hora almacenada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al almacenar la hora"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/registerAnswer": {
            "post": {
                "tags": [
                    "Respuestas"
                ],
                "summary": "Registro final de las respuestas",
                "description": "Registra las respuestas de un cliente",
                "operationId": "registerAnswer",
                "parameters": [
                    {
                        "name": "Authorization_",
                        "in": "header",
                        "description": "Bearer token for authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "JSON data to import",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "json": {
                                        "type": "string",
                                        "example": "{\"question1\":\"edwin@gmail.com\",\"question2\":\"Item 1\",\"question3\":\"2024-06-21T15:39\",\"question4\":\"Image 2\"}"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Respuesta almacenada correctamente"
                    },
                    "400": {
                        "description": "Invalid JSON data supplied"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Log in a user",
                "description": "Loguea un usuario admin.",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "User log in",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "user@prueba.com",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid email/password supplied"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "description": "Registra un usuario Admin.",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "User registration",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "user@prueba.com",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "User registered successfully"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/begin": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Legea o registra un usuario",
                "description": "Atravez de la primera respuesta la cual es el correo electronico se logea o registra un usuario devuelve el token de seguridad para el resto de la encuesta",
                "operationId": "160f38f17729a45d91fc236aa139fa70",
                "requestBody": {
                    "description": "Cliente login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "survey": {
                                        "description": "Nombre de la encuesta",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "user@prueba.com",
                                    "survey": "EncuentadepruebaparaNeurociencia"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Respuesta registrada exitosamente"
                    },
                    "400": {
                        "description": "Error al registrar la respuesta"
                    }
                }
            }
        },
        "/importFromJson": {
            "post": {
                "tags": [
                    "Administrador de encuestas"
                ],
                "summary": "Import data from JSON - IMPORTA ENCUESTA DESDE JSON",
                "description": "Para importar el contenido de una encuesta desde un JSON convertido a string, para ello debe antes eliminar las referencia a imágenes, además debe está logueado como administrador",
                "operationId": "importFromJson",
                "parameters": [
                    {
                        "name": "Authorization_",
                        "in": "header",
                        "description": "Bearer token for authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "JSON data to import",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "json": {
                                        "type": "string",
                                        "example": "{\"title\":\"Encuesta de prueba\",\"description\":\"Encuesta de prueba\",\"pages\":[{\"name\":\"page1\",\"elements\":[{\"type\":\"dropdown\",\"name\":\"question1\",\"title\":\"¿Cuál es tu color favorito?\",\"choices\":[{\"value\":\"Rojo\",\"text\":\"Rojo\"},{\"value\":\"Azul\",\"text\":\"Azul\"},{\"value\":\"Verde\",\"text\":\"Verde\"}]},{\"type\":\"text\",\"name\":\"question2\",\"title\":\"¿Cuál es tu correo electrónico?\",\"inputType\":\"email\"},{\"type\":\"imagepicker\",\"name\":\"question3\",\"title\":\"¿Cuál es tu animal favorito?\",\"choices\":[{\"value\":\"Perro\",\"text\":\"Perro\"},{\"value\":\"Gato\",\"text\":\"Gato\"},{\"value\":\"Pájaro\",\"text\":\"Pájaro\"}]}]}]}"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data imported successfully"
                    },
                    "400": {
                        "description": "Invalid JSON data supplied"
                    }
                }
            }
        }
    }
}